Fixes bug with localized PayPal EC button
The button URL for the EC button was using WPLANG to
dynamically populate the locale code using the WP
language setting.  If this value was empty, though,
the URL was invalid was resulting in a broken image.

This was fixed by checking for a valid value first,
and if no value is present it falls to en_US by default